tests: Move to new class
